Clean up LiftoffAction
