--- lib/actions/abstract_action.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-import '../stores/accounts_store.dart';
-
-@immutable
-abstract class LiftoffAction {
-  String get name;
-  String get tooltip;
-  IconData get icon;
-  Color get activeColor;
-  bool get isActivated;
-
-  Future<void> Function(UserData userData) get invoke;
-}
