+++ lib/widgets/post/post.dart
@@ -3,8 +3,8 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:lemmy_api_client/v3.dart';
 import 'package:nested/nested.dart';
 
-import '../../actions/post.dart';
 import '../../hooks/logged_in_action.dart';
+import '../../liftoff_action.dart';
 import '../../pages/full_post/full_post.dart';
 import '../../stores/config_store.dart';
 import '../../util/async_store_listener.dart';
@@ -93,7 +93,7 @@ class _Post extends HookWidget {
         builder: (context, store) => WithSwipeActions(
           actions: [
             PostUpvoteAction(post: postStore, context: context),
-            PostSaveAction(post: postStore, context: context),
+            PostSaveAction(post: postStore),
           ],
           onTrigger: (action) => loggedInAction(action.invoke)(),
           child: DecoratedBox(
