+++ lib/widgets/swipe_actions.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
-import '../actions/abstract_action.dart';
+import '../liftoff_action.dart';
 import 'post/post.dart';
 
 /// Widget that wraps [child] and allows for swipe actions to be performed on it.
